backend/requirements.txt
# Python dependencies for Mavn Bench backend

# FastAPI and async
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
aiofiles==23.2.1
asyncpg==0.29.0

# Pydantic and validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Skip for now - Python 3.13 compatibility issue

# Redis
redis==5.0.1

# OpenTelemetry
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-jaeger==1.21.0

# Document processing
PyMuPDF==1.23.8  # PDF processing
python-docx==1.1.0  # Word documents
openpyxl==3.1.2  # Excel files
beautifulsoup4==4.12.2  # Web scraping
playwright==1.40.0  # JS-heavy sites
yt-dlp==2023.11.16  # YouTube/podcast download

# Search engines
qdrant-client==1.7.0  # Vector search
neo4j==5.14.0  # Graph database
elasticsearch==8.11.0  # Full-text search

# Utilities
python-multipart==0.0.6  # File uploads
pyyaml==6.0.1  # YAML config
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-dateutil==2.8.2  # Date utilities

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
